PRMDR-473 refactor bulk upload report handler
